table_selection_to_mime_data: Move to utils
