+++ Orange/widgets/utils/tableview.py
@@ -1,4 +1,8 @@
-from AnyQt.QtCore import Signal, QItemSelectionModel, Qt, QSize, QEvent
+import io
+import csv
+
+from AnyQt.QtCore import Signal, QItemSelectionModel, Qt, QSize, QEvent, \
+    QByteArray, QMimeData
 from AnyQt.QtGui import QMouseEvent
 from AnyQt.QtWidgets import QTableView, QStyleOptionViewItem, QStyle
 
@@ -77,3 +81,44 @@ class TableView(QTableView):
         if event.type() in (QEvent.StyleChange, QEvent.FontChange):
             table_view_compact(self)
         super().changeEvent(event)
+
+
+def table_selection_to_mime_data(table):
+    """Copy the current selection in a QTableView to the clipboard.
+    """
+    lines = table_selection_to_list(table)
+
+    as_csv = lines_to_csv_string(lines, dialect="excel").encode("utf-8")
+    as_tsv = lines_to_csv_string(lines, dialect="excel-tab").encode("utf-8")
+
+    mime = QMimeData()
+    mime.setData("text/csv", QByteArray(as_csv))
+    mime.setData("text/tab-separated-values", QByteArray(as_tsv))
+    mime.setData("text/plain", QByteArray(as_tsv))
+    return mime
+
+
+def lines_to_csv_string(lines, dialect="excel"):
+    stream = io.StringIO()
+    writer = csv.writer(stream, dialect=dialect)
+    writer.writerows(lines)
+    return stream.getvalue()
+
+
+def table_selection_to_list(table):
+    model = table.model()
+    indexes = table.selectedIndexes()
+
+    rows = sorted(set(index.row() for index in indexes))
+    columns = sorted(set(index.column() for index in indexes))
+
+    lines = []
+    for row in rows:
+        line = []
+        for col in columns:
+            val = model.index(row, col).data(Qt.DisplayRole)
+            # TODO: use style item delegate displayText?
+            line.append(str(val))
+        lines.append(line)
+
+    return lines
